Incread signing speed of manytimessig and statelessig
The previous OTS scheme used a single-word checksum of base 65536
which turned out to be the main bottleneck when signing and generating
the keys. Signing this checksum would take about as long as
signing the rest of the message.

A speedup was achieved by replacing this checksum by 2 words of base
256. This nearly doubled the signing speed of the one-time signatures
and schemes that rely on them (manytimesig and statelessig) The
downside of this is that the size of a one-time signature was
increased from 1056 to 1088 bytes. for the manytimessig scheme this
means an increased signature size of 5219 bytes for the manytimessig
scheme (from 5123 bytes)